Former-commit-id: 5612c8c32c78ffa7c6d290de4eb99bcb2020b904

+++ Tools/Launcher/Classes/appscommandssender.h
@@ -17,6 +17,6 @@ public:
     bool RequestQuit(const QString& appPath);
 
 private:
-    long SendMessage(long message, const QString& appPath);
+    int SendMessage(int message, const QString& appPath);
     QLocalSocket* socket = nullptr;
 };
